Language.CNF.Parse.ParseDIMACS

Description

A simple module for parsing CNF files in DIMACS format.

Documentation

parseByteString :: SourceName -> ByteString -> Either ParseError CNF

Parse a byte string containing DIMACS CNF data. The source name is only | used in error messages and may be the empty string.

parseFile :: FilePath -> IO (Either ParseError CNF)

Parse a file containing DIMACS CNF data.

data CNF

Constructors

CNF

Fields

numVars :: !Int

The number of variables in the problem as reported by the cnf header.

numClauses :: !Int

The number of clauses in the problem as reported by the cnf header.

clauses :: ![Clause]

type Clause = UArray Int Int